pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-it-ops"
version = "0.0.1"
description = "MCP server for IT ops: query system health, Grafana alerts, freqtrade bots, and more from Claude."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Bastiaan Rusch", email = "bastiaan@bastiaan365.nl" }]
keywords = ["mcp", "claude", "homelab", "monitoring", "it-ops"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "mcp>=1.27.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
mcp-it-ops = "mcp_it_ops.server:main"

[project.urls]
Homepage = "https://github.com/bastiaan365/mcp-it-ops"
Issues = "https://github.com/bastiaan365/mcp-it-ops/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
